release: v0.9——进行中的今天按半天计;小组件三行等权重
计天:
- 缺的那半天补捕窗口还没关时(晚点还没到 / 还没过 12 点),单样本只计 0.5 天,
  晚点打上后自动补满;窗口已关(如 20 点才有晚点)仍按整天,与过去的日子一致
- 今天不再算「无记录」催补记;PunchRules.pendingSlots 判定窗口,双端同一套规则
- YearStats.recordedDays 改 Double(= 各城市天数之和),首页「天已记录」可显示 .5
- 首页今日卡加一行说明「今天先算半天 · 晚点打上后补满一天」

小组件:
- Top 1 不再是 44pt 锚点、2/3 名不再是脚注:年份眉题 + Top 3 三行等权重
  (城市名 15 / 天数 19–20 / 单位 11,同字重同色),整块唯一品牌色移到年份
- 行间与末尾弹性间距平分余量,三行均匀铺开;Android 按启动器竖屏高度
  (OPTION_APPWIDGET_MAX_HEIGHT)显示 3 / 2 / 1 行,行隐藏时同时收掉其前面的间距
- :core 测试 62 → 67(进行中的今天 4 个场景 + 补捕窗口)

+++ android/core/src/main/kotlin/app/terndays/core/Models.kt
@@ -59,7 +59,8 @@ data class YearStats(
     val year: Int,
     val firstDate: LocalDate,
     val lastDate: LocalDate,
-    val recordedDays: Int,
+    /** 已记录天数（= 各城市天数之和；跨城日 0.5+0.5，进行中的今天可能是 0.5） */
+    val recordedDays: Double,
     val cities: List<CityStat>,
     val unrecordedDates: List<LocalDate>,
     val days: Map<LocalDate, DayAttribution>,
